Add validation for git metadata env vars
