+++ dd-java-agent/agent-bootstrap/src/test/groovy/datadog/trace/bootstrap/instrumentation/decorator/TestDecoratorTest.groovy
@@ -4,7 +4,8 @@ import datadog.trace.api.DDTags
 import datadog.trace.api.sampling.PrioritySampling
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan
 import datadog.trace.bootstrap.instrumentation.api.Tags
-import datadog.trace.bootstrap.instrumentation.ci.CIProviderInfo
+import datadog.trace.bootstrap.instrumentation.ci.CIInfo
+import datadog.trace.bootstrap.instrumentation.ci.CITagsProvider
 import datadog.trace.bootstrap.instrumentation.ci.git.GitInfo
 
 class TestDecoratorTest extends BaseDecoratorTest {
@@ -83,16 +84,16 @@ class TestDecoratorTest extends BaseDecoratorTest {
   }
 
   def newMockCiInfo() {
-    return new CIProviderInfo() {
+    return new CITagsProvider() {
 
         @Override
-        protected GitInfo buildCIGitInfo() {
+        GitInfo buildCIGitInfo() {
           return GitInfo.NOOP
         }
 
         @Override
-        protected CIProviderInfo.CIInfo buildCIInfo() {
-          return CIProviderInfo.CIInfo.NOOP
+        CIInfo buildCIInfo() {
+          return CIInfo.NOOP
         }
 
         @Override
