+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/ci/CIInfo.java
@@ -1,0 +1,251 @@
+package datadog.trace.bootstrap.instrumentation.ci;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+
+public class CIInfo {
+  public static final CIInfo NOOP = new CIInfo();
+
+  public static Builder builder() {
+    return new Builder();
+  }
+
+  public static final class Builder {
+    private String ciProviderName;
+    private String ciPipelineId;
+    private String ciPipelineName;
+    private String ciStageName;
+    private String ciJobName;
+    private String ciPipelineNumber;
+    private String ciPipelineUrl;
+    private String ciJobUrl;
+    private String ciWorkspace;
+    private Map<String, String> ciEnvVars;
+
+    public Builder ciProviderName(String ciProviderName) {
+      this.ciProviderName = ciProviderName;
+      return this;
+    }
+
+    public Builder ciPipelineId(String ciPipelineId) {
+      this.ciPipelineId = ciPipelineId;
+      return this;
+    }
+
+    public Builder ciPipelineName(String ciPipelineName) {
+      this.ciPipelineName = ciPipelineName;
+      return this;
+    }
+
+    public Builder ciStageName(String ciStageName) {
+      this.ciStageName = ciStageName;
+      return this;
+    }
+
+    public Builder ciJobName(String ciJobName) {
+      this.ciJobName = ciJobName;
+      return this;
+    }
+
+    public Builder ciPipelineNumber(String ciPipelineNumber) {
+      this.ciPipelineNumber = ciPipelineNumber;
+      return this;
+    }
+
+    public Builder ciPipelineUrl(String ciPipelineUrl) {
+      this.ciPipelineUrl = ciPipelineUrl;
+      return this;
+    }
+
+    public Builder ciJobUrl(String ciJobUrl) {
+      this.ciJobUrl = ciJobUrl;
+      return this;
+    }
+
+    public Builder ciWorkspace(String ciWorkspace) {
+      this.ciWorkspace = ciWorkspace;
+      return this;
+    }
+
+    public Builder ciEnvVars(String... ciEnvVarKeysArray) {
+      if (ciEnvVarKeysArray == null || ciEnvVarKeysArray.length == 0) {
+        return this;
+      }
+
+      ciEnvVars = new HashMap<>();
+      for (String ciEnvVarKey : ciEnvVarKeysArray) {
+        final String envVarVal = System.getenv(ciEnvVarKey);
+        if (envVarVal != null && !envVarVal.isEmpty()) {
+          ciEnvVars.put(ciEnvVarKey, envVarVal);
+        }
+      }
+      return this;
+    }
+
+    public CIInfo build() {
+      return new CIInfo(
+          ciProviderName,
+          ciPipelineId,
+          ciPipelineName,
+          ciStageName,
+          ciJobName,
+          ciPipelineNumber,
+          ciPipelineUrl,
+          ciJobUrl,
+          ciWorkspace,
+          ciEnvVars);
+    }
+  }
+
+  private final String ciProviderName;
+  private final String ciPipelineId;
+  private final String ciPipelineName;
+  private final String ciStageName;
+  private final String ciJobName;
+  private final String ciPipelineNumber;
+  private final String ciPipelineUrl;
+  private final String ciJobUrl;
+  private final String ciWorkspace;
+  private final Map<String, String> ciEnvVars;
+
+  public CIInfo() {
+    this(null, null, null, null, null, null, null, null, null, null);
+  }
+
+  public CIInfo(
+      String ciProviderName,
+      String ciPipelineId,
+      String ciPipelineName,
+      String ciStageName,
+      String ciJobName,
+      String ciPipelineNumber,
+      String ciPipelineUrl,
+      String ciJobUrl,
+      String ciWorkspace,
+      Map<String, String> ciEnvVars) {
+    this.ciProviderName = ciProviderName;
+    this.ciPipelineId = ciPipelineId;
+    this.ciPipelineName = ciPipelineName;
+    this.ciStageName = ciStageName;
+    this.ciJobName = ciJobName;
+    this.ciPipelineNumber = ciPipelineNumber;
+    this.ciPipelineUrl = ciPipelineUrl;
+    this.ciJobUrl = ciJobUrl;
+    this.ciWorkspace = ciWorkspace;
+    this.ciEnvVars = ciEnvVars;
+  }
+
+  public String getCiProviderName() {
+    return ciProviderName;
+  }
+
+  public String getCiPipelineId() {
+    return ciPipelineId;
+  }
+
+  public String getCiPipelineName() {
+    return ciPipelineName;
+  }
+
+  public String getCiStageName() {
+    return ciStageName;
+  }
+
+  public String getCiJobName() {
+    return ciJobName;
+  }
+
+  public String getCiPipelineNumber() {
+    return ciPipelineNumber;
+  }
+
+  public String getCiPipelineUrl() {
+    return ciPipelineUrl;
+  }
+
+  public String getCiJobUrl() {
+    return ciJobUrl;
+  }
+
+  public String getCiWorkspace() {
+    return ciWorkspace;
+  }
+
+  public Map<String, String> getCiEnvVars() {
+    return ciEnvVars;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    CIInfo ciInfo = (CIInfo) o;
+    return Objects.equals(ciProviderName, ciInfo.ciProviderName)
+        && Objects.equals(ciPipelineId, ciInfo.ciPipelineId)
+        && Objects.equals(ciPipelineName, ciInfo.ciPipelineName)
+        && Objects.equals(ciStageName, ciInfo.ciStageName)
+        && Objects.equals(ciJobName, ciInfo.ciJobName)
+        && Objects.equals(ciPipelineNumber, ciInfo.ciPipelineNumber)
+        && Objects.equals(ciPipelineUrl, ciInfo.ciPipelineUrl)
+        && Objects.equals(ciJobUrl, ciInfo.ciJobUrl)
+        && Objects.equals(ciWorkspace, ciInfo.ciWorkspace)
+        && Objects.equals(ciEnvVars, ciInfo.ciEnvVars);
+  }
+
+  @Override
+  public int hashCode() {
+    int hash = 1;
+    hash = 31 * hash + (ciProviderName == null ? 0 : ciProviderName.hashCode());
+    hash = 31 * hash + (ciPipelineId == null ? 0 : ciPipelineId.hashCode());
+    hash = 31 * hash + (ciPipelineName == null ? 0 : ciPipelineName.hashCode());
+    hash = 31 * hash + (ciStageName == null ? 0 : ciStageName.hashCode());
+    hash = 31 * hash + (ciJobName == null ? 0 : ciJobName.hashCode());
+    hash = 31 * hash + (ciPipelineNumber == null ? 0 : ciPipelineNumber.hashCode());
+    hash = 31 * hash + (ciPipelineUrl == null ? 0 : ciPipelineUrl.hashCode());
+    hash = 31 * hash + (ciJobUrl == null ? 0 : ciJobUrl.hashCode());
+    hash = 31 * hash + (ciWorkspace == null ? 0 : ciWorkspace.hashCode());
+    hash = 31 * hash + (ciEnvVars == null ? 0 : ciEnvVars.hashCode());
+    return hash;
+  }
+
+  @Override
+  public String toString() {
+    return "CIInfo{"
+        + "ciProviderName='"
+        + ciProviderName
+        + '\''
+        + ", ciPipelineId='"
+        + ciPipelineId
+        + '\''
+        + ", ciPipelineName='"
+        + ciPipelineName
+        + '\''
+        + ", ciStageName='"
+        + ciStageName
+        + '\''
+        + ", ciJobName='"
+        + ciJobName
+        + '\''
+        + ", ciPipelineNumber='"
+        + ciPipelineNumber
+        + '\''
+        + ", ciPipelineUrl='"
+        + ciPipelineUrl
+        + '\''
+        + ", ciJobUrl='"
+        + ciJobUrl
+        + '\''
+        + ", ciWorkspace='"
+        + ciWorkspace
+        + '\''
+        + ", ciEnvVars='"
+        + ciEnvVars
+        + '\''
+        + '}';
+  }
+}
