+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/ci/utils/PathUtils.java
@@ -1,0 +1,17 @@
+package datadog.trace.bootstrap.instrumentation.ci.utils;
+
+public class PathUtils {
+  public static String expandTilde(final String path) {
+    if (path == null || !path.startsWith("~")) {
+      return path;
+    }
+
+    if (!path.equals("~") && !path.startsWith("~/")) {
+      // Home dir expansion is not supported for other user.
+      // Returning path without modifications.
+      return path;
+    }
+
+    return path.replaceFirst("^~", System.getProperty("user.home"));
+  }
+}
