+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/GithubActionsInfoTest.groovy
@@ -4,12 +4,7 @@ class GithubActionsInfoTest extends CIProviderInfoTest {
 
   @Override
   CIProviderInfo instanceProvider() {
-    return new GithubActionsInfo() {
-        @Override
-        protected String getGitFolderName() {
-          return GIT_FOLDER_FOR_TESTS
-        }
-      }
+    return new GithubActionsInfo()
   }
 
   @Override
