+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/JenkinsInfoTest.groovy
@@ -4,12 +4,7 @@ class JenkinsInfoTest extends CIProviderInfoTest {
 
   @Override
   CIProviderInfo instanceProvider() {
-    return new JenkinsInfo() {
-        @Override
-        protected String getGitFolderName() {
-          return GIT_FOLDER_FOR_TESTS
-        }
-      }
+    return new JenkinsInfo()
   }
 
   @Override
