+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/TravisInfoTest.groovy
@@ -4,12 +4,7 @@ class TravisInfoTest extends CIProviderInfoTest {
 
   @Override
   CIProviderInfo instanceProvider() {
-    return new TravisInfo() {
-        @Override
-        protected String getGitFolderName() {
-          return GIT_FOLDER_FOR_TESTS
-        }
-      }
+    return new TravisInfo()
   }
 
   @Override
