+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/ci/git/GitUtilsTest.groovy
@@ -14,10 +14,28 @@ class GitUtilsTest extends DDSpecification {
     result == expectedPersonInfo
 
     where:
-    author | expectedPersonInfo
-    null | PersonInfo.NOOP
-    ""|PersonInfo.NOOP
-    "wrong-data"|PersonInfo.NOOP
-    "Jane Doe <jane.doe@email.com>"| janeDoePersonInfo
+    author                          | expectedPersonInfo
+    null                            | PersonInfo.NOOP
+    ""                              | PersonInfo.NOOP
+    "wrong-data"                    | PersonInfo.NOOP
+    "Jane Doe <jane.doe@email.com>" | janeDoePersonInfo
+  }
+
+  def "test commit SHA validity (#sha): #expectedResult "() {
+    when:
+    def result = GitUtils.isValidCommitSha(sha)
+
+    then:
+    result == expectedResult
+
+    where:
+    sha                                        | expectedResult
+    null                                       | false
+    ""                                         | false
+    "123456789"                                | false
+    "1234567890123456789012345678901234567890" | true
+    "1234567890123456789012345678901234abcdef" | true
+    "1234567890123456789012345678901234ABCDEF" | true
+    "1234567890123456789012345678901234ABCDEX" | false
   }
 }
